+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 * @mynameisstephen fixed an issue where slider event listeners were not being cleaned up ([view](https://github.com/videojs/video.js/pull/1475))
 * @alexrqs cleaned up the Spanish translation ([view](https://github.com/videojs/video.js/pull/1494))
 * @t2y added a Japanese translation ([view](https://github.com/videojs/video.js/pull/1497))
+* @chikathreesix fixed an issue where data-setup options could be missed ([view](https://github.com/videojs/video.js/pull/1514))
 
 --------------------